backend/requirements.txt
fastapi
uvicorn
python-multipart
python-docx
pydantic
openai
sqlalchemy
striprtf
pymupdf
fpdf2
